Migrate pdf endpoint to report service
