+++ product_code/build.sbt
@@ -137,10 +137,9 @@ lazy val matriculation_service_api = (project in file("matriculation_service/api
 lazy val matriculation_service = (project in file("matriculation_service/impl"))
   .enablePlugins(LagomScala)
   .settings(libraryDependencies += lagomScaladslKafkaBroker)
-  .settings(libraryDependencies += Dependencies.pdf)
   .settings(Settings.implSettings("matriculation_service"))
   .dependsOn(user_service_api % withTests, certificate_service_api % withTests, examreg_service_api % withTests, operation_service_api % withTests,
-    shared_server % withTests, shared_client, matriculation_service_api, hyperledger_component_impl, pdf_processing_api)
+    shared_server % withTests, shared_client, matriculation_service_api, hyperledger_component_impl)
 
 lazy val certificate_service_api = (project in file("certificate_service/api"))
   .settings(Settings.apiSettings("certificate_service_api"))
@@ -235,9 +234,10 @@ lazy val report_service = (project in file("report_service/impl"))
   .settings(
     libraryDependencies ++= Dependencies.implDefaultDependencies,
     libraryDependencies ++= Dependencies.defaultPersistenceKafkaDependencies,
-    libraryDependencies += Dependencies.zip
+    libraryDependencies += Dependencies.zip,
+    libraryDependencies ++= Dependencies.pdfSigning
   )
   .settings(Settings.implSettings("report_service"))
   .dependsOn(report_service_api % withTests, admission_service_api % withTests,
     user_service_api % withTests, certificate_service_api % withTests, matriculation_service_api % withTests, course_service_api % withTests,
-    shared_client % withTests, shared_server % withTests)
+    shared_client % withTests, shared_server % withTests, pdf_processing_api)
