+++ product_code/matriculation_service/impl/src/main/resources/prod-application.conf
@@ -1,28 +1,16 @@
 include "application"
 
-uc4 {
-    hyperledger {
-        walletPath = "/opt/docker/share/hyperledger/wallet_path"
-        networkConfig = "/var/libs/configmaps/connection_profile_kubernetes.yaml"
-        tlsCert = "/var/secrets/org/cert.pem"
-        username = ${HYPERLEDGER_USERNAME}
-        password = ${HYPERLEDGER_PASSWORD}
-        organisationId = ${HYPERLEDGER_ORGANISATION_ID}
-        organisationName = ${HYPERLEDGER_ORGANISATION_NAME}
-        channel = ${HYPERLEDGER_CHANNEL}
-        chaincode = ${HYPERLEDGER_CHAINCODE}
-        caURL = ${HYPERLEDGER_CA_URL}
-    }
-
-    pdf {
-        keyStorePath = ${?KEYSTORE_PATH}
-        certificateEnrollmentHtml = ${?CERTIFICATE_ENROLLMENT_PATH}
-        keyStorePassword = ${?KEYSTORE_PASSWORD}
-        certificateAlias = ${?CERTIFICATE_ALIAS}
-        tsaURL = ${?TSA_URL}
-        address = ${?ADDRESS}
-        organization = ${?ORGANIZATION}
-    }
+uc4.hyperledger {
+  walletPath = "/opt/docker/share/hyperledger/wallet_path"
+  networkConfig = "/var/libs/configmaps/connection_profile_kubernetes.yaml"
+  tlsCert = "/var/secrets/org/cert.pem"
+  username = ${HYPERLEDGER_USERNAME}
+  password = ${HYPERLEDGER_PASSWORD}
+  organisationId = ${HYPERLEDGER_ORGANISATION_ID}
+  organisationName = ${HYPERLEDGER_ORGANISATION_NAME}
+  channel = ${HYPERLEDGER_CHANNEL}
+  chaincode = ${HYPERLEDGER_CHAINCODE}
+  caURL = ${HYPERLEDGER_CA_URL}
 }
 
 play {
