+++ product_code/report_service/impl/src/main/resources/application.conf
@@ -14,6 +14,16 @@ uc4 {
     }
   }
 
+  pdf {
+    keyStorePath = "/keystore.jks"
+    keyStorePassword = "upbuc4"
+    certificateAlias = "key"
+    tsaURL = "http://timestamp.digicert.com/"
+    certificateEnrollmentHtml = "certificateEnrollment.html"
+    address = "University Credits 4.0\nPaderborn University\nWarburger Straße 100\n33098 Paderborn\nGermany\n"
+    organization = "University Credits 4.0"
+  }
+
   # timeouts in milliseconds
   timeouts {
     validation = 500
