+++ product_code/report_service/impl/src/main/resources/prod-application.conf
@@ -8,11 +8,23 @@ play {
   http.secret.key = ${APPLICATION_SECRET}
 }
 
-uc4.secrets {
-  master = ${UC4_MASTER_SECRET}
-  salts {
-    kafka = ${UC4_KAFKA_SALT}
-  }
+uc4 {
+    secrets {
+      master = ${UC4_MASTER_SECRET}
+      salts {
+        kafka = ${UC4_KAFKA_SALT}
+      }
+    }
+
+    pdf {
+      keyStorePath = ${?KEYSTORE_PATH}
+      certificateEnrollmentHtml = ${?CERTIFICATE_ENROLLMENT_PATH}
+      keyStorePassword = ${?KEYSTORE_PASSWORD}
+      certificateAlias = ${?CERTIFICATE_ALIAS}
+      tsaURL = ${?TSA_URL}
+      address = ${?ADDRESS}
+      organization = ${?ORGANIZATION}
+    }
 }
 
 db.default {
