+++ product_code/report_service/impl/src/main/scala/de/upb/cs/uc4/report/impl/signature/SigningService.java
@@ -1,4 +1,4 @@
-package de.upb.cs.uc4.matriculation.impl.signature;
+package de.upb.cs.uc4.report.impl.signature;
 
 import org.apache.commons.io.FileUtils;
 import org.apache.pdfbox.pdmodel.PDDocument;
